requirements.txt
flask
gunicorn
tensorflow==2.18.0
pillow
numpy
opencv-python-headless
requests
